requirements.txt
asgiref==3.3.1
Django==3.1.7
django-extensions==3.1.1
django-jsonfield==1.4.1
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
postgis==1.0.4
psycopg2==2.8.6
PyJWT==2.0.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
